Specify a UID/GID for vagrant user in kickstart.
GID_MIN changed from 500 in RHEL6 to 1000 in RHEL7.  This ends up
causing a clash with the "isunix" account if the host is added to Puppet
due to VirtualBox Guest Additions for 5.2.6 creating a "vboxadd" user
with GID of 1001 (even though it doesn't use the group, it gets put into
the GID 1 "bin" group, thanks Oracle.)

Creating the vagrant user with GID of GID_MIN + 2 clears the way for
"isunix" to be added later.