Initial commit of Interim RHEL 8 build.
This build depends on users manually registering with the Red Hat
Customer Portal in order to obtain updates using their RMIT Red Hat
login credentials.

This will be changed back to registration based on Activation Key
shortly, which will bring it back in line with the RHEL 6/7 builds and
should allow for an easy conversion to Satellite once the new Satellite
server is available for RHEL 8 use.