Remove unnecessary code regarding journald.
So I apparently got fixated enough on the problem of how to compare two
floating point numbers in a shell script that I didn't stop to think
that it was totally unnecessary, given that RHEL 6 doesn't use systemd
and thus doesn't have journald.  The adjustment will only ever be
applicable for RHEL 7+, so why not just put the necessary config in the
relevant kickstart.cfg file?

Still, it was an interesting concept to learn. :P