Disable auto-update for vagrant-vbguest plugin.
If the vagrant-vbguest plugin is detected auto-update will be set to
"false".  This is required for our build because the plugin fires before
the box is fully brought up with hostname, networking, and the part of
provisioning that registers the box with Satellite.  If vbguest tries to
update the VirtualBox Guest Additions and needs to install or update
packages prior to registration with Satellite, the process will fail.

To re-enable auto-update functionality add "config.vbguest.auto_update
= true" to your Vagrantfile.