Convert config to build off Satellite 6.
In order for this to work, various environment variables need to be set
prior to running 'make'. For example, for RHEL7:

- VAGRANT_RHEL7_ORG="<orgid>"
- VAGRANT_RHEL7_KEY="<activationkey>'

It is recommended that this be done via a file that can be sourced,
especially if running 'build-all'.

Note that there is a difference between the way Satellite 5 and 6 deal
with unregistering systems.  The old 'satellite-deregister' script
deleted the system totally from Satellite 5, the new script for
Satellite 6 does not; it only deregisters them.  As a result deleting
the system profile entirely is currently a manual process.  Details are
in the README.md file.

Also note that the RHEL6 build currently installs the EPEL repo during
the build as some of the default packages come via EPEL and the new
Satellite does not currently have them available.

Additional notes:

- Updates the RHEL7 build to 7.7
- Build ISOs are still hosted on the old Satellite 5.8 server, until
  such time as we replace it with another "general repo" server
- Adds a "clean-tempfiles" target to the Makefile now that files are
  built from templates.  This also removes the output directory if it
  exists.